[+] feature -> /!\ improving the RoundAreaCommand with multiple shapes !
The goal is to calculate the area that a shape would require, as a round.

However, we quickly see that the command only calculate the round area for rounds !
We would like to use the RoundUtil to calculate the round area required by a quadrilateral shape;
thus avoiding having calculation methods duplicated.

We'll see how adapter can help us do that in the next commit!